Fixed debounce issue. Typo in the pin read function. I had the port reference where the pin reference should have been.

+++ STM32-Code/Synth_Clock_Module/Core/Inc/Structures.h
@@ -81,7 +81,7 @@ typedef struct TempoButton
 	volatile Button button;
 	volatile uint8_t tap_count;
 	volatile uint8_t tap_interval_buffer_size;
-	volatile uint16_t tap_intervals[];
+	volatile uint32_t tap_intervals[];
 }TempoButton;
